레이아웃 1차 정규화

+++ Common Library/DotNet_Standard/PosePacket/Service/Auth/P_E_TokenRefresh.cs
@@ -3,6 +3,6 @@
     public class O_TokenRefresh
     {
         public string PoseToken { get; set; }
-        public int TokenExpireIn { get; set; }
+        public long TokenExpireIn { get; set; }
     }
 }
